I have added 3 types of errors.
Class CityDataValidation - handles errors when a city is not present in the database.
Class CityNameProcessor - handles errors when a city name is entered, and its last letter is not valid.
Class CityNameValidator - checks if the name is written in the Ukrainian language.
And the class ExceptionTest tests all the added errors.